move maybe_truncate into run_context
